requirements.txt
Django<1.8
beanstalkc
PyYAML
django-model-utils
django-extensions
django-transaction-signals
South>=1.0.0